Allow chat window reuse
Allow users to set modify the plugin behavior so that a copilot chat
window is reused instead of re-creating chats every time.

By 'g:copilot_reuse_active_chat' to 1, (default is 0), the plugin will
check if there's an active chat buffer upon access attempt, and switch
to it instead of opening a new chat window.

If the buffer exists but isn't viewable in the current window, the
plugin will create a new window split and open the active chat window
there.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>