Unset active chat when deleting it
When deleting a copilot chat buffer with `:bdelete`, it leaves the
buffer in a limbo state, where one can still access the buffer, but the
buffer doesn't function anymore as an active copilot chat, so make sure
it isn't marked as the active chat window anymore.

Note that the optimal solution should be a 'post-delete' event (that
doesn't exist yet), wiping out the buffer altogether using 'bwipeout',
but for, this should do.

Another possible solution is resurrect the active buffer if it still
exists.

The downside is a user might create many chat windows, delete them, thus
creating many zombie chat windows, but this is true for original state
anyway.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>