Manage zombie buffers wipeout
A continuation for previous patch. When deleting a chat window buffer,
other than setting the active window buffer to -1 (if the deleted buffer
was the active one), also remember which buffer was delete using a
special variable.

On the next buffer delete, wipeout the zombie buffer before putting the
newly delete buffer number as the zombie variable value. Thus, a user
may delete as many chat buffers as he wishes, while keeping a maximum of
1 zombie buffer in the background.

Again, it would have been better to just use a post-delete kind of
event, but this should be  good enough. Another possibility is to have
background async thread of some sort to check the status of zombie chat
buffers and reap them once their deletion is completed, but this is
a slight overkill for such (not-really) an issue.

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>