Configuration revamp
Create a proper generic API of configuration, completely independent
than other layers of the plugin. Allowing a layered structure of the
different subsystems in the code.

The basic ideas is as follows:

The configuration API provides a couple of function (putting aside the
data directory creation function). One for getting a value and one for
setting one. All configuration clients should use this functionality to
handle their configuration settings. This approach puts an end to all
problems with configuration access problems, as the get/set reads and
writes the configuration file as needed, for their client should even be
aware of the file and its structure.

For example, the models selection is now a client of the new
configuration API. Not interacting directly with the configuration, but
calling an API for getting and setting configuration values.
(See further commit for models/configuration separation).

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>